requirements.txt
auth0-python
celery
dj-database-url
Django==1.11.6
django-colorful
djangorestframework
gunicorn
oauthlib
psycopg2
pycrypto
python-dotenv
python-jose
python-openid
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
vine
whitenoise
